+++ resources/defineSamplingWindow.py
@@ -316,17 +316,6 @@ class defineSamplingWindow(QWidget):
         self.interactive_plot.fig.canvas.draw()
         self.interactive_plot.fig.canvas.setFocus()
 
-    #---------------------------------------------------------------------------------------------
-    def sync_with_item(self, item):
-        if item != self.item: return
-
-        self.raise_()
-
-        xlim = self.interactive_plot.axs[0].get_xlim()
-        ylim = self.interactive_plot.axs[0].get_ylim()
-        self.interactive_plot.axs[0].clear()
-        self.myplot(limits=[xlim,ylim])
-
     #----------------------------------------------------------------------------------
     @staticmethod
     def sampling(series, sampling_index, kind="linear", integrated=False, ax=None, quad_points=20):